use new shared function for restoring main window
