+++ src/node/desktop/src/main/preferences/electron-desktop-options.ts
@@ -14,7 +14,7 @@
  *
  */
 
-import { BrowserWindow, screen } from 'electron';
+import { BrowserWindow } from 'electron';
 import Store from 'electron-store';
 import { existsSync, lstatSync } from 'fs';
 import { basename, dirname, join } from 'path';
@@ -26,7 +26,7 @@ import { RStudioUserState } from '../../types/user-state-schema';
 import { generateSchema, legacyPreferenceManager } from './../preferences/preferences';
 import DesktopOptions from './desktop-options';
 import { kWindowsRExe } from '../../ui/utils';
-import { WindowBounds, intersects } from '../window-utils';
+import { WindowBounds, positionAndEnsureVisible } from '../window-utils';
 
 const kProportionalFont = 'font.proportionalFont';
 const kFixedWidthFont = 'font.fixedWidthFont';
@@ -162,45 +162,9 @@ export class DesktopOptionsImpl implements DesktopOptions {
   private restoreMainWindowBoundsImpl(mainWindow: BrowserWindow): void {
     const savedBounds = this.windowBounds();
 
-    // Check if saved bounds is still partially in one of the available displays.
-
-    // Shrink the window rectangle a bit just to capture cases like RStudio
-    // too close to edge of window and hardly showing at all.
-    const checkRect = {
-      height: savedBounds.height - 5,
-      width: savedBounds.width - 5,
-      x: savedBounds.x + 5,
-      y: savedBounds.y + 5,
-    };
-
-    // check for intersection
-    const goodDisplays = screen.getAllDisplays().find((display) => {
-      return intersects(checkRect, display.workArea);
-    });
-
-    // Restore it to previous location if possible, or center of primary display otherwise
-    if (goodDisplays) {
-      mainWindow.setBounds(savedBounds);
-    } else {
-      const primaryBounds = screen.getPrimaryDisplay().bounds;
-      const newSize = {
-        width: Math.min(properties.view.default.windowBounds.width, primaryBounds.width),
-        height: Math.min(properties.view.default.windowBounds.height, primaryBounds.height),
-      };
-
-      mainWindow.setSize(newSize.width, newSize.height);
-
-      // window.center() doesn't consistently pick the primary display,
-      // so manually calculating the center of the primary display
-      mainWindow.setPosition(
-        primaryBounds.x + (primaryBounds.width - newSize.width) / 2,
-        primaryBounds.y + (primaryBounds.height - newSize.height) / 2,
-      );
-    }
-
-    // ensure a minimum size for the window on restore
-    const currSize = mainWindow.getSize();
-    mainWindow.setSize(Math.max(300, currSize[0]), Math.max(200, currSize[1]));
+    positionAndEnsureVisible(
+      mainWindow, savedBounds,
+      properties.view.default.windowBounds.width, properties.view.default.windowBounds.height);
 
     if (savedBounds.maximized) {
       mainWindow.maximize();
